+++ packages/assets/src/cache/index.ts
@@ -1,0 +1,1 @@
+export * from './Cache';
