+++ packages/assets/src/index.ts
@@ -1,1 +1,5 @@
 export * from './Assets';
+export * from './utils';
+export * from './cache';
+export * from './loader';
+export * from './resolver';
