+++ packages/assets/src/loader/types.ts
@@ -1,0 +1,13 @@
+import { LoaderParser } from './parsers';
+
+export interface LoadAsset<T=any>
+{
+    src: string;
+    data?: T;
+}
+
+export interface PromiseAndParser
+{
+    promise: Promise<any>
+    parser: LoaderParser
+}
