+++ packages/assets/src/resolver/parsers/index.ts
@@ -1,0 +1,2 @@
+export * from './spriteSheetUrlParser';
+export * from './textureUrlParser';
