+++ packages/assets/src/resolver/parsers/textureUrlParser.ts
@@ -1,7 +1,7 @@
 import { settings } from '@pixi/settings';
 
-import { loadTextures } from '../loader';
-import { ResolveAsset } from '../resolver/Resolver';
+import { loadTextures } from '../../loader';
+import { ResolveAsset } from '../types';
 
 export const textureUrlParser = {
     test: loadTextures.test,
