+++ packages/assets/src/utils/detections/index.ts
@@ -1,0 +1,2 @@
+export * from './detectAvif';
+export * from './detectWebp';
