+++ packages/assets/src/utils/url/isAbsoluteUrl.ts
